# docs/api/admin.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "admin-service",
    "version": "0.1.0"
  },
  "paths": {
...
    "/admin/costs/daily": {
      "get": {
        "tags": [
          "costs"
        ],
        "summary": "Get Daily Costs",
        "description": "Aggregate LLM + vendor API costs across the last ``days`` days.",
        "operationId": "get_daily_costs_admin_costs_daily_get",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 365,
              "minimum": 1,
              "default": 30,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyCostsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/costs/by-agent": {
      "get": {
        "tags": [
          "costs"
        ],
        "summary": "Get Costs By Agent",
        "description": "Aggregate ``model_runs`` cost per agent for one calendar month.",
        "operationId": "get_costs_by_agent_admin_costs_by_agent_get",
        "parameters": [
          {
            "name": "month",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Calendar month in YYYY-MM.",
              "title": "Month"
            },
            "description": "Calendar month in YYYY-MM."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CostsByAgentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentCostEntry": {
        "properties": {
          "agent_id": {
            "type": "string",
            "title": "Agent Id"
          },
          "runs": {
            "type": "integer",
            "title": "Runs"
          },
          "model_runs": {
            "type": "integer",
            "title": "Model Runs"
          },
          "input_tokens": {
            "type": "integer",
            "title": "Input Tokens"
          },
          "output_tokens": {
            "type": "integer",
            "title": "Output Tokens"
          },
          "cost_usd": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Cost Usd"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "agent_id",
          "runs",
          "model_runs",
          "input_tokens",
          "output_tokens",
          "cost_usd"
        ],
        "title": "AgentCostEntry",
        "description": "One agent's cost rollup for a calendar month."
      },
...
      "CostsByAgentResponse": {
        "properties": {
          "month": {
            "type": "string",
            "title": "Month"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "title": "End Date"
          },
          "agents": {
            "items": {
              "$ref": "#/components/schemas/AgentCostEntry"
            },
            "type": "array",
            "title": "Agents"
          },
          "total_cost_usd": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Cost Usd"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "month",
          "start_date",
          "end_date",
          "agents",
          "total_cost_usd"
        ],
        "title": "CostsByAgentResponse",
        "description": "``GET /admin/costs/by-agent`` payload."
      },
      "DailyCostEntry": {
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "model_cost_usd": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Model Cost Usd"
          },
          "api_cost_usd": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Api Cost Usd"
          },
          "total_cost_usd": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Cost Usd"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "date",
          "model_cost_usd",
          "api_cost_usd",
          "total_cost_usd"
        ],
        "title": "DailyCostEntry",
        "description": "One day of aggregated LLM + vendor API costs."
      },
      "DailyCostsResponse": {
        "properties": {
          "days": {
            "type": "integer",
            "title": "Days"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "title": "End Date"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/DailyCostEntry"
            },
            "type": "array",
            "title": "Entries"
          },
          "total_cost_usd": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Cost Usd"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "days",
          "start_date",
          "end_date",
          "entries",
          "total_cost_usd"
        ],
        "title": "DailyCostsResponse",
        "description": "``GET /admin/costs/daily`` payload."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}